requirements.txt
crewai
fastapi
python-dotenv
pydantic
uvicorn